# backend/services/product-service/generated/openapi.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.2 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// Defines values for SearchProductsParamsSortBy.
const (
	CreatedAt SearchProductsParamsSortBy = "created_at"
	Price     SearchProductsParamsSortBy = "price"
	Rating    SearchProductsParamsSortBy = "rating"
)

// Defines values for SearchProductsParamsSortOrder.
const (
	Asc  SearchProductsParamsSortOrder = "asc"
	Desc SearchProductsParamsSortOrder = "desc"
)

// Product defines model for Product.
type Product struct {
	// Category カテゴリ
	Category *string `json:"category,omitempty"`

	// CreatedAt 作成日時
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// Description 説明
	Description *string `json:"description,omitempty"`

	// Id 商品ID
	Id string `json:"id"`

	// Images 画像URL
	Images *[]string `json:"images,omitempty"`

	// IsActive アクティブ状態
	IsActive *bool `json:"is_active,omitempty"`

	// IsSale セール中か
	IsSale *bool `json:"is_sale,omitempty"`

	// Name 商品名
	Name string `json:"name"`

	// Price 価格
	Price float64 `json:"price"`

	// Rating 評価
	Rating *float64 `json:"rating,omitempty"`

	// ReviewCount レビュー数
	ReviewCount *int `json:"review_count,omitempty"`

	// SalePrice セール価格
	SalePrice *float64 `json:"sale_price"`

	// Stock 在庫数
	Stock *int `json:"stock,omitempty"`

	// UpdatedAt 更新日時
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// ProductWithSKUs defines model for ProductWithSKUs.
type ProductWithSKUs struct {
	// Category カテゴリ
	Category *string `json:"category,omitempty"`

	// CreatedAt 作成日時
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// Description 説明
	Description *string `json:"description,omitempty"`

	// Id 商品ID
	Id string `json:"id"`

	// Images 画像URL
	Images *[]string `json:"images,omitempty"`

	// IsActive アクティブ状態
	IsActive *bool `json:"is_active,omitempty"`

	// IsSale セール中か
	IsSale *bool `json:"is_sale,omitempty"`

	// Name 商品名
	Name string `json:"name"`

	// Price 基本価格
	Price float64 `json:"price"`

	// Rating 評価
	Rating *float64 `json:"rating,omitempty"`

	// ReviewCount レビュー数
	ReviewCount *int `json:"review_count,omitempty"`

	// SalePrice セール価格
	SalePrice *float64 `json:"sale_price"`

	// Skus SKU一覧
	Skus []SKU `json:"skus"`

	// UpdatedAt 更新日時
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// SKU defines model for SKU.
type SKU struct {
	// Attributes バリエーション属性（色、サイズなど）
	Attributes *map[string]string `json:"attributes,omitempty"`

	// CreatedAt 作成日時
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// Id SKU ID
	Id string `json:"id"`

	// IsActive アクティブ状態
	IsActive *bool `json:"is_active,omitempty"`

	// Name SKU名
	Name string `json:"name"`

	// Price SKU価格
	Price float64 `json:"price"`

	// ProductId 商品ID
	ProductId string `json:"product_id"`

	// SkuCode SKUコード
	SkuCode string `json:"sku_code"`

	// Stock 在庫数
	Stock int `json:"stock"`

	// UpdatedAt 更新日時
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// SearchResult defines model for SearchResult.
type SearchResult struct {
	// CurrentPage 現在ページ
	CurrentPage *int `json:"current_page,omitempty"`

	// PageSize ページサイズ
	PageSize *int       `json:"page_size,omitempty"`
	Products *[]Product `json:"products,omitempty"`

	// Total 総件数
	Total *int `json:"total,omitempty"`

	// TotalPages 総ページ数
	TotalPages *int `json:"total_pages,omitempty"`
}

// GetProductsByCategoryParams defines parameters for GetProductsByCategory.
type GetProductsByCategoryParams struct {
	// Page ページ番号
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// PageSize ページサイズ
	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`
}

// SearchProductsParams defines parameters for SearchProducts.
type SearchProductsParams struct {
	// Keyword 検索キーワード
	Keyword *string `form:"keyword,omitempty" json:"keyword,omitempty"`

	// Category カテゴリ
	Category *string `form:"category,omitempty" json:"category,omitempty"`

	// PriceMin 最低価格
	PriceMin *float32 `form:"price_min,omitempty" json:"price_min,omitempty"`

	// PriceMax 最高価格
	PriceMax *float32 `form:"price_max,omitempty" json:"price_max,omitempty"`

	// SortBy ソート対象
	SortBy *SearchProductsParamsSortBy `form:"sort_by,omitempty" json:"sort_by,omitempty"`

	// SortOrder ソート順序
	SortOrder *SearchProductsParamsSortOrder `form:"sort_order,omitempty" json:"sort_order,omitempty"`

	// Page ページ番号
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// PageSize ページサイズ
	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`
}

// SearchProductsParamsSortBy defines parameters for SearchProducts.
type SearchProductsParamsSortBy string

// SearchProductsParamsSortOrder defines parameters for SearchProducts.
type SearchProductsParamsSortOrder string

// GetLowStockSKUsParams defines parameters for GetLowStockSKUs.
type GetLowStockSKUsParams struct {
	// Threshold 在庫閾値
	Threshold *int `form:"threshold,omitempty" json:"threshold,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// カテゴリー一覧取得
	// (GET /products/categories)
	GetCategories(ctx echo.Context) error
	// カテゴリ別商品取得
	// (GET /products/category/{category})
	GetProductsByCategory(ctx echo.Context, category string, params GetProductsByCategoryParams) error
	// おすすめ商品取得
	// (GET /products/featured)
	GetFeaturedProducts(ctx echo.Context) error
	// 商品検索
	// (GET /products/search)
	SearchProducts(ctx echo.Context, params SearchProductsParams) error
	// 商品詳細取得
	// (GET /products/{id})
	GetProduct(ctx echo.Context, id string) error
	// 商品のSKU一覧取得
	// (GET /products/{product_id}/skus)
	GetSKUsByProduct(ctx echo.Context, productId string) error
	// 商品とSKUをまとめて取得
	// (GET /products/{product_id}/with-skus)
	GetProductWithSKUs(ctx echo.Context, productId string) error
	// SKUコード検索
	// (GET /skus/code/{code})
	GetSKUByCode(ctx echo.Context, code string) error
	// 低在庫SKU一覧
	// (GET /skus/low-stock)
	GetLowStockSKUs(ctx echo.Context, params GetLowStockSKUsParams) error
	// SKU詳細取得
	// (GET /skus/{id})
	GetSKU(ctx echo.Context, id string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetCategories converts echo context to params.
func (w *ServerInterfaceWrapper) GetCategories(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetCategories(ctx)
	return err
}

// GetProductsByCategory converts echo context to params.
func (w *ServerInterfaceWrapper) GetProductsByCategory(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "category" -------------
	var category string

	err = runtime.BindStyledParameterWithLocation("simple", false, "category", runtime.ParamLocationPath, ctx.Param("category"), &category)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter category: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetProductsByCategoryParams
	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", ctx.QueryParams(), &params.Page)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter page: %s", err))
	}

	// ------------- Optional query parameter "page_size" -------------

	err = runtime.BindQueryParameter("form", true, false, "page_size", ctx.QueryParams(), &params.PageSize)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter page_size: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetProductsByCategory(ctx, category, params)
	return err
}

// GetFeaturedProducts converts echo context to params.
func (w *ServerInterfaceWrapper) GetFeaturedProducts(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetFeaturedProducts(ctx)
	return err
}

// SearchProducts converts echo context to params.
func (w *ServerInterfaceWrapper) SearchProducts(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params SearchProductsParams
	// ------------- Optional query parameter "keyword" -------------

	err = runtime.BindQueryParameter("form", true, false, "keyword", ctx.QueryParams(), &params.Keyword)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter keyword: %s", err))
	}

	// ------------- Optional query parameter "category" -------------

	err = runtime.BindQueryParameter("form", true, false, "category", ctx.QueryParams(), &params.Category)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter category: %s", err))
	}

	// ------------- Optional query parameter "price_min" -------------

	err = runtime.BindQueryParameter("form", true, false, "price_min", ctx.QueryParams(), &params.PriceMin)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter price_min: %s", err))
	}

	// ------------- Optional query parameter "price_max" -------------

	err = runtime.BindQueryParameter("form", true, false, "price_max", ctx.QueryParams(), &params.PriceMax)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter price_max: %s", err))
	}

	// ------------- Optional query parameter "sort_by" -------------

	err = runtime.BindQueryParameter("form", true, false, "sort_by", ctx.QueryParams(), &params.SortBy)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter sort_by: %s", err))
	}

	// ------------- Optional query parameter "sort_order" -------------

	err = runtime.BindQueryParameter("form", true, false, "sort_order", ctx.QueryParams(), &params.SortOrder)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter sort_order: %s", err))
	}

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", ctx.QueryParams(), &params.Page)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter page: %s", err))
	}

	// ------------- Optional query parameter "page_size" -------------

	err = runtime.BindQueryParameter("form", true, false, "page_size", ctx.QueryParams(), &params.PageSize)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter page_size: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SearchProducts(ctx, params)
	return err
}

// GetProduct converts echo context to params.
func (w *ServerInterfaceWrapper) GetProduct(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetProduct(ctx, id)
	return err
}

// GetSKUsByProduct converts echo context to params.
func (w *ServerInterfaceWrapper) GetSKUsByProduct(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "product_id" -------------
	var productId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "product_id", runtime.ParamLocationPath, ctx.Param("product_id"), &productId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter product_id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSKUsByProduct(ctx, productId)
	return err
}

// GetProductWithSKUs converts echo context to params.
func (w *ServerInterfaceWrapper) GetProductWithSKUs(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "product_id" -------------
	var productId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "product_id", runtime.ParamLocationPath, ctx.Param("product_id"), &productId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter product_id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetProductWithSKUs(ctx, productId)
	return err
}

// GetSKUByCode converts echo context to params.
func (w *ServerInterfaceWrapper) GetSKUByCode(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "code" -------------
	var code string

	err = runtime.BindStyledParameterWithLocation("simple", false, "code", runtime.ParamLocationPath, ctx.Param("code"), &code)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter code: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSKUByCode(ctx, code)
	return err
}

// GetLowStockSKUs converts echo context to params.
func (w *ServerInterfaceWrapper) GetLowStockSKUs(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetLowStockSKUsParams
	// ------------- Optional query parameter "threshold" -------------

	err = runtime.BindQueryParameter("form", true, false, "threshold", ctx.QueryParams(), &params.Threshold)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter threshold: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetLowStockSKUs(ctx, params)
	return err
}

// GetSKU converts echo context to params.
func (w *ServerInterfaceWrapper) GetSKU(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSKU(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/products/categories", wrapper.GetCategories)
	router.GET(baseURL+"/products/category/:category", wrapper.GetProductsByCategory)
	router.GET(baseURL+"/products/featured", wrapper.GetFeaturedProducts)
	router.GET(baseURL+"/products/search", wrapper.SearchProducts)
	router.GET(baseURL+"/products/:id", wrapper.GetProduct)
	router.GET(baseURL+"/products/:product_id/skus", wrapper.GetSKUsByProduct)
	router.GET(baseURL+"/products/:product_id/with-skus", wrapper.GetProductWithSKUs)
	router.GET(baseURL+"/skus/code/:code", wrapper.GetSKUByCode)
	router.GET(baseURL+"/skus/low-stock", wrapper.GetLowStockSKUs)
	router.GET(baseURL+"/skus/:id", wrapper.GetSKU)

}